+++ chess/src/main/java/pdp/model/ai/heuristics/Heuristic.java
@@ -3,5 +3,5 @@ package pdp.model.ai.heuristics;
 import pdp.model.board.Board;
 
 public interface Heuristic {
-  int evaluate(Board board, boolean isWhite);
+  float evaluate(Board board, boolean isWhite);
 }
